+++ CHANGELOG.md
@@ -49,7 +49,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Include table:
   - A mechanism to set the font, using a text box (#748)
   - `force_font_width` to force a font width (#752)
-- PCB Print: a mechanism to filter components for a particular layer (#706)
+- PCB Print:
+  - A mechanism to filter components for a particular layer (#706)
+  - Workaround for Ghostscript handling %d wrong when followed by an hex digit
+    (#763)
 - Report:
   - `top_total`, `bot_total`, `total_smd`, `total_tht` and `total_all`
     component counts (See #730)