+++ docs/source/Changelog.rst
@@ -82,8 +82,12 @@ Added
    -  A mechanism to set the font, using a text box (#748)
    -  ``force_font_width`` to force a font width (#752)
 
--  PCB Print: a mechanism to filter components for a particular layer
-   (#706)
+-  PCB Print:
+
+   -  A mechanism to filter components for a particular layer (#706)
+   -  Workaround for Ghostscript handling %d wrong when followed by an
+      hex digit (#763)
+
 -  Report:
 
    -  ``top_total``, ``bot_total``, ``total_smd``, ``total_tht`` and